+++ logics/main_logics.py
@@ -29,16 +29,6 @@ openai_api_key = os.getenv('OPENAI_API_KEY')
 client = OpenAI(api_key=openai_api_key)
 
 
-def artdata_to_string(data):
-    """예술품 관련 data를 string 형태로 변환
-    Args:
-        data (dict): 예술품 관련 data
-    """
-    ret_str = ""
-    for each_key in data.keys():
-        ret_str += "[" + each_key + "]: " + data[each_key]
-    return ret_str
-
 
 def get_art_data_from_db(query, db_art):
     """query와 관련한 DB의 data를 검색해서 string 형태로 도출
@@ -55,12 +45,7 @@ def get_art_data_from_db(query, db_art):
     art_idx.add(np.array(db_art_emb))
     art_dists, art_idxs = art_idx.search(np.array([query_embed]), 3)
     
-
-    data_string = ""
-    data_string += "예술품 관련 data 목록: ["
-    for i in art_idxs[0]:
-        data_string += "예술품 data" + str(i + 1) + ": " + artdata_to_string(db_art[i][0]) + "//"
-    data_string += "]"
+    data_string = main_prompts.make_db_art_to_string(db_art, art_idxs)
     
     return data_string
 
@@ -79,12 +64,8 @@ def get_etc_data_from_db(query, db_etc):
     db_etc_emb = [each_data[1] for each_data in db_etc]
     etc_idx.add(np.array(db_etc_emb))
     etc_dists, etc_idxs = etc_idx.search(np.array([query_embed]), 7)
-
-    data_string = ""
-    data_string += "기타 다양한 data 목록: ["
-    for i in etc_idxs[0]:
-        data_string += "기타 data" + str(i + 1) + ": " + db_etc[i][0] + "//"
-    data_string += "]"
+    
+    data_string = main_prompts.make_db_etc_to_string(db_etc, etc_idxs)
 
     return data_string
